test: move the attachHeaders test to main_test.go
The test exercises attachHeaders(), which is defined in main.go. Keep it
next to the other main.go tests.
