+++ src/config_test.go
@@ -3,11 +3,9 @@ package src
 import (
 	"context"
 	"net/http"
-	"net/http/httptest"
 	"testing"
 
 	"github.com/sevensolutions/traefik-oidc-auth/src/config"
-	"github.com/sevensolutions/traefik-oidc-auth/src/session"
 )
 
 func newTestConfig() *config.Config {
@@ -68,50 +66,5 @@ func TestNew_RejectsInvalidHeaderTemplate(t *testing.T) {
 
 	if _, err := New(context.Background(), http.NotFoundHandler(), cfg, "test"); err == nil {
 		t.Fatal("expected an invalid header template to be rejected")
-	}
-}
-
-func TestAttachHeaders_UsesThePreparsedTemplates(t *testing.T) {
-	cfg := newTestConfig()
-	cfg.Headers = []config.HeaderConfig{
-		{Name: "X-Subject", Value: "{{ .claims.sub }}"},
-		{Name: "X-Roles", Values: "{{ .claims.roles | mapToJsonArray }}"},
-		{Name: "X-Broken-Values", Values: "not a json array"},
-		{Name: "X-Empty"},
-	}
-
-	handler, err := New(context.Background(), http.NotFoundHandler(), cfg, "test")
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	toa := handler.(*TraefikOidcAuth)
-
-	req := httptest.NewRequest("GET", "https://app.example.com/", nil)
-	req.Header.Set("X-Broken-Values", "spoofed")
-
-	claims := map[string]interface{}{
-		"sub":   "user-1",
-		"roles": []interface{}{"admin", "user"},
-	}
-
-	if err := toa.attachHeaders(req, &session.SessionState{}, claims, false, true); err != nil {
-		t.Fatal(err)
-	}
-
-	if got := req.Header.Get("X-Subject"); got != "user-1" {
-		t.Errorf("expected the rendered subject, got %q", got)
-	}
-
-	if got := req.Header.Values("X-Roles"); len(got) != 2 || got[0] != "admin" || got[1] != "user" {
-		t.Errorf("expected both roles as separate values, got %v", got)
-	}
-
-	if got := req.Header.Values("X-Broken-Values"); len(got) != 0 {
-		t.Errorf("expected a header whose values fail to render to be removed, got %v", got)
-	}
-
-	if _, ok := req.Header["X-Empty"]; !ok {
-		t.Error("expected a header without a value to still be set")
 	}
 }
