+++ src/main_test.go
@@ -2,8 +2,14 @@ package src
 
 import (
 	"bytes"
+	"context"
 	"encoding/json"
+	"net/http"
+	"net/http/httptest"
 	"testing"
+
+	"github.com/sevensolutions/traefik-oidc-auth/src/config"
+	"github.com/sevensolutions/traefik-oidc-auth/src/session"
 )
 
 func TestTemplate_mapToJsonArray(t *testing.T) {
@@ -43,5 +49,50 @@ func TestTemplate_mapToJsonArray(t *testing.T) {
 
 	if result[2] != "prefix:123:suffix" {
 		t.Errorf("Expected prefix:123:suffix at index 2, got %s", result[2])
+	}
+}
+
+func TestAttachHeaders_UsesThePreparsedTemplates(t *testing.T) {
+	cfg := newTestConfig()
+	cfg.Headers = []config.HeaderConfig{
+		{Name: "X-Subject", Value: "{{ .claims.sub }}"},
+		{Name: "X-Roles", Values: "{{ .claims.roles | mapToJsonArray }}"},
+		{Name: "X-Broken-Values", Values: "not a json array"},
+		{Name: "X-Empty"},
+	}
+
+	handler, err := New(context.Background(), http.NotFoundHandler(), cfg, "test")
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	toa := handler.(*TraefikOidcAuth)
+
+	req := httptest.NewRequest("GET", "https://app.example.com/", nil)
+	req.Header.Set("X-Broken-Values", "spoofed")
+
+	claims := map[string]interface{}{
+		"sub":   "user-1",
+		"roles": []interface{}{"admin", "user"},
+	}
+
+	if err := toa.attachHeaders(req, &session.SessionState{}, claims, false, true); err != nil {
+		t.Fatal(err)
+	}
+
+	if got := req.Header.Get("X-Subject"); got != "user-1" {
+		t.Errorf("expected the rendered subject, got %q", got)
+	}
+
+	if got := req.Header.Values("X-Roles"); len(got) != 2 || got[0] != "admin" || got[1] != "user" {
+		t.Errorf("expected both roles as separate values, got %v", got)
+	}
+
+	if got := req.Header.Values("X-Broken-Values"); len(got) != 0 {
+		t.Errorf("expected a header whose values fail to render to be removed, got %v", got)
+	}
+
+	if _, ok := req.Header["X-Empty"]; !ok {
+		t.Error("expected a header without a value to still be set")
 	}
 }
